More refactorings
Processor with >1.5k loc was getting hard to manage. Split the logic
into individual files.
