+++ lib/scala/interpreter-dsl/src/main/java/org/enso/interpreter/dsl/builtins/ClassName.java
@@ -1,0 +1,7 @@
+package org.enso.interpreter.dsl.builtins;
+
+public record ClassName(String pkg, String name) {
+    public String fullyQualifiedName() {
+        return pkg + "." + name;
+    }
+}
