+++ lib/scala/interpreter-dsl/src/main/java/org/enso/interpreter/dsl/builtins/TypeKind.java
@@ -1,0 +1,7 @@
+package org.enso.interpreter.dsl.builtins;
+
+public enum TypeKind {
+    VOID,
+    ARRAY,
+    OBJECT
+}
